+++ src/components/top-up-field/asset-amount-input/styles.tsx
@@ -1,6 +1,6 @@
-import { transparent } from '../../../../config/styles';
-import { createUseStyles } from '../../../../styles/create-use-styles';
-import { formatSize } from '../../../../styles/format-size';
+import { transparent } from 'src/config/styles';
+import { createUseStyles } from 'src/styles/create-use-styles';
+import { formatSize } from 'src/styles/format-size';
 
 export const useTopUpAssetAmountInputStyles = createUseStyles(({ colors, typography }) => ({
   headerContainer: {
