+++ src/components/top-up-field/fiat-icons/cny.png
