+++ src/components/top-up-field/fiat-icons/egp.png
