+++ src/components/top-up-field/fiat-icons/kwd.png
